load-test/requirements.txt
faker
psycopg2-binary==2.9.10
prometheus_client==0.20.0